Moving example JS out of src/js to src/explorer
