+++ src/explorer/js/deviceMenu.js
